+++ matchms/Spectrum.py
@@ -182,10 +182,9 @@ class Spectrum:
         if "fingerprint" in spectrum_dict:
             spectrum_dict["fingerprint"] = spectrum_dict["fingerprint"].tolist()
         return spectrum_dict
-    
-    def update_keys(self, key_conversions: dict):
-        self._metadata.update_keys(key_conversions)
-        return self
+
+    def metadata_dict(self, export_style: str = "matchms") -> dict:
+        return self._metadata.to_dict(export_style)
 
     @property
     def mz(self):