+++ tests/test_metadata.py
@@ -55,6 +55,21 @@ def test_metadata_setitem_getitem(harmonize, set_key, set_value, get_key, get_va
         "Expected different _metadata dictionary."
 
 
+@pytest.mark.parametrize("input_dict, export_style, expected", [
+    [None, "matchms", {}],
+    [{"precursor_mz": 101.01}, "matchms", {"precursor_mz": 101.01}],
+    [{"peptide_modifications": 1}, "massbank", {"COMMENT:PEPTIDE_MODIFICATIONS": 1}],
+    [{"ionmode": "Negative"}, "massbank", {"AC$MASS_SPECTROMETRY:ION_MODE": "Negative"}],
+    [{"compound_name": "Dummy"}, "nist", {"Name": "Dummy"}],
+    [{"compound_name": "Dummy"}, "riken", {"NAME": "Dummy"}],
+    [{"compound_name": "Dummy"}, "gnps", {"NAME": "Dummy"}],
+    ])
+def test_metadata_to_dict(input_dict, export_style, expected):
+    metadata = Metadata(input_dict)
+    assert metadata.to_dict(export_style) == expected, \
+        "Expected different metadata dictionary."
+
+
 @pytest.mark.parametrize("dict1, dict2, expected", [
     [{"precursor_mz": 101.01}, {"precursor_mz": 101.01}, True],
     [{"abc": ["a", 5, 7.01]}, {"abc": ["a", 7.01, 5]}, False],